Bug #354311 use unsigned value to calculate nsec
[bug number  ] 354311
[root cause  ] singed value right shift cause higher bit is 1
[changes     ] use unsigned value to calculate nsec
[side effects] no
[self test   ] <sp9630ea><pass>
[whether AOB ] no
[reviewers   ] geng.ren;xiongshan.an;jiajun.wu;

+++ kernel/time/timekeeping.c
@@ -162,7 +162,7 @@ static inline s64 timekeeping_get_ns(struct timekeeper *tk)
 {
 	cycle_t cycle_now, cycle_delta;
 	struct clocksource *clock;
-	s64 nsec;
+	u64 nsec;
 
 	/* read clocksource: */
 	clock = tk->clock;
@@ -175,7 +175,7 @@ static inline s64 timekeeping_get_ns(struct timekeeper *tk)
 	nsec >>= tk->shift;
 
 	/* If arch requires, add in get_arch_timeoffset() */
-	return nsec + get_arch_timeoffset();
+	return (s64)(nsec + get_arch_timeoffset());
 }
 
 static inline s64 timekeeping_get_ns_raw(struct timekeeper *tk)
